+++ package.json
@@ -10,6 +10,8 @@
     "// Generates LavaMoat secured modules from the Webpack build": "",
     "build:lavamoat": "lavamoat ./scripts/lavamoat-browserify.js --policy ./lavamoat/build-lavamoat/policy.json --override ./lavamoat/build-lavamoat/policy-override.json",
     "dev": "webpack --config ./webpack.config.dev.js --watch",
+    "playground": "PLAYGROUND=default webpack --config ./webpack.config.dev.js --watch",
+    "playground:ds": "PLAYGROUND=ds webpack --config ./webpack.config.dev.js --watch",
     "lint": "eslint . --cache",
     "// Generates LavaMoat policies for everything that is protected w/ LavaMoat": "",
     "policy": "yarn policy:build:webpack && yarn policy:build:lavamoat && yarn policy:bundle",
@@ -38,9 +40,9 @@
     "@vanilla-extract/sprinkles": "1.5.0",
     "chroma-js": "2.4.2",
     "clsx": "1.2.1",
-    "lodash": "4.17.21",
     "ethers": "5.7.1",
     "eventemitter3": "4.0.7",
+    "lodash": "4.17.21",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "wagmi": "0.6.8"
@@ -51,6 +53,7 @@
     "@types/chroma-js": "2.1.4",
     "@types/chrome": "0.0.197",
     "@types/lodash": "4.14.186",
+    "@types/node": "18.11.0",
     "@types/react": "18.0.21",
     "@types/react-dom": "18.0.6",
     "@typescript-eslint/eslint-plugin": "5.38.1",
