+++ src/entries/popup/global.css.ts
@@ -3,6 +3,6 @@ import { globalStyle } from '@vanilla-extract/css';
 globalStyle('html, body', {
   margin: 0,
   padding: 0,
-  width: '320px',
-  height: '420px',
+  width: '360px',
+  height: '600px',
 });
