+++ src/entries/popup/pages/_playgrounds/index.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { Default } from './default';
+import { DesignSystem } from './ds';
+
+export const PlaygroundComponents = {
+  default: <Default />,
+  ds: <DesignSystem />,
+};
