+++ tsconfig.json
@@ -19,5 +19,5 @@
       "~/*": ["src/*"]
     },
   },
-  "include": ["src"],
+  "include": ["src", ".env.d.ts"],
 }
